refactor(core): extract workspace federation read model
